# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/payments/queue:
    get:
      operationId: listPaymentVerificationQueue
      tags: [admin]
      summary: List orders awaiting payment verification (oldest first), with a signed proof URL
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated verification queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerificationQueue"
  /admin/orders/{id}/approve:
    post:
      operationId: approveOrderPayment
      tags: [admin]
      summary: Approve a payment — moves the order to pagado and decrements stock (idempotent)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Order not awaiting verification, or insufficient stock to fulfil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/orders/{id}/reject:
    post:
      operationId: rejectOrderPayment
      tags: [admin]
      summary: Reject a payment — moves the order to rechazado (idempotent, no stock change)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Order not awaiting verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/shipments:
    get:
      operationId: listShipments
      tags: [admin]
      summary: List paid orders in fulfilment (logistics board), filterable by method and status
      parameters:
        - name: deliveryMethod
          in: query
          required: false
          schema:
            type: string
            enum: [delivery, recojo]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [en_preparacion, enviado, entregado, recojo_pendiente, recogido, cancelado]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Paginated shipments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShipmentList"
  /admin/orders/{id}/fulfillment:
    post:
      operationId: advanceFulfillment
      tags: [admin]
      summary: Advance an order's fulfilment status (validated state transition)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvanceFulfillmentInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Order not in fulfilment, or invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List all products for inventory management (includes inactive)
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated admin products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductList"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product (optionally with occasions and variants)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "409":
          description: Slug or SKU already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products/{id}:
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product's fields (partial); soft-delete via active=false
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products/{id}/variants:
    post:
      operationId: createVariant
      tags: [admin]
      summary: Add a size/color variant to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariantInput"
      responses:
        "200":
          description: Updated product (with the new variant)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: SKU or size/color combination already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products/{id}/media:
    post:
      operationId: attachProductMedia
      tags: [admin]
      summary: Attach an uploaded photo/video to a product (appended last)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachProductMediaInput"
      responses:
        "200":
          description: Updated product (with the new media)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/media/{id}:
    delete:
      operationId: deleteProductMedia
      tags: [admin]
      summary: Remove a product media item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Media not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/products/import:
    post:
      operationId: importProducts
      tags: [admin]
      summary: Bulk import products/variants from CSV (upsert by SKU); dryRun previews errors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductImportInput"
      responses:
        "200":
          description: Import result (row-level errors, imported count)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImportResult"
        "400":
          description: CSV could not be parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/variants/{id}:
    patch:
      operationId: updateVariant
      tags: [admin]
      summary: Update a variant (stock, price override, active, size/color/sku)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariantInput"
      responses:
        "200":
          description: Updated product (owning the variant)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProduct"
        "404":
          description: Variant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: SKU or size/color combination already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/reports:
    get:
      operationId: getSalesReport
      tags: [admin]
      summary: Sales, funnel and abandoned-cart report for a date range (admin only)
      parameters:
        - name: from
          in: query
          required: false
          description: ISO date YYYY-MM-DD (inclusive). Defaults to 30 days ago.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
        - name: to
          in: query
          required: false
          description: ISO date YYYY-MM-DD (inclusive). Defaults to today.
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /admin/coupons:
    get:
      operationId: listCoupons
      tags: [admin]
      summary: List coupons for management (admin only)
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated coupons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponList"
    post:
      operationId: createCoupon
      tags: [admin]
      summary: Create a coupon (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponInput"
      responses:
        "201":
          description: Created coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "409":
          description: Coupon code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /admin/coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [admin]
      summary: Update a coupon's fields (partial); activate/deactivate via active
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCouponInput"
      responses:
        "200":
          description: Updated coupon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
        "404":
          description: Coupon not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Coupon code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteCoupon
      tags: [admin]
      summary: Delete a coupon (admin only); blocked if already redeemed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "404":
          description: Coupon not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Coupon already redeemed, cannot delete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /admin/config/media/upload-url:
    post:
      operationId: createConfigMediaUploadUrl
      tags: [admin]
      summary: Signed URL to upload a QR/banner image to public Storage (admin only)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigMediaUploadInput"
      responses:
        "200":
          description: Signed upload URL plus the public read URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigMediaUploadUrl"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    PaymentVerificationQueue:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PaymentVerificationItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
...
    ShipmentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShipmentItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    AdvanceFulfillmentInput:
      type: object
      properties:
        to:
          type: string
          enum: [en_preparacion, enviado, entregado, recojo_pendiente, recogido, cancelado]
      required:
        - to
...
    AdminProductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminProduct"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
...